Move to model.
